# Improve Qt object management and performance